#beh_manager: move WorldStateMock
