+++ core/server/ServerCore.test.js
@@ -1,22 +1,7 @@
 const ServerCore = require("./ServerCore");
-const WorldState = require("../state/WorldState");
-const CharFactory = require("../state/CharFactory");
+const WorldStateMock = require("../tests/WorldStateMock");
 const expect = require("expect");
 
-class WorldStateMock extends WorldState {
-    loadSave() {
-        const realmUnits = [
-            CharFactory.initEmptyCharacter({
-                wishes: [
-                    { name: "FollowWish", targetUnitId: this.uniqueId() },
-                ],
-            }),
-        ];
-        this.state.units = realmUnits;
-
-        return Promise.resolve();
-    }
-}
 
 describe("ServerCore", () => {
     describe("with wishManager", () => {
